Added Properties (WIP)
This commit adds the basis of GLib property support to the bindings. The
user can currently set properties for a subset of types, and the bindings
appear to behave appropriately.

An html/javascript utility called 'converter.html' was added to generate
C# PInvoke/FuncLoader signatures from plain C signatures. It is a very
rudimentary implementation.

In order to faciliate this, the following types have been added:
 - GType (Type system integration between C# and GLib)
    => This supports associating a GType with a System.Type and being
       able to retrieve the corresponding type/compare types using
       casting and operator overloads.
 - GValue (Generic values)
    => This supports initialising values and setting them. Getters and
       all other functionality is WIP and very broken.

Property getters/setters can be implemented using a very simple API:

```c#
public int DefaultHeight
{
    get => GetProperty<int>("default-height");
    set => SetProperty<int>("default-height", value);
}
```

The current implementation is very messy and I'm surprised it actually
works as-is. Next steps are to finish the ability to get properties
(presently only setting properties is supported), as well as to clean
up the code for GValue.